Extract filter logic into download/filter.rs
Move asset filtering, dedup/collision detection, and version
selection out of download/mod.rs (6258 lines) into a dedicated
filter module. Pure code move with visibility adjustments — no
behavior changes.

Moved: NormalizedPath, determine_media_type, apply_raw_policy,
extract_skip_candidates, version_with_fallback, pre_ensure_asset_dir,
filter_asset_to_tasks, DownloadTask, and ~90 associated tests.

mod.rs: 6258 → 3953 lines, filter.rs: 2414 lines.
